A tracked link follows the URL a browser would: the href is entity-decoded before it is signed
An href reads "?utm_source=email&amp;utm_medium=email" in HTML, and the
rewriter signed that literal string, so the redirect landed with "&amp;"
in the query and every parameter after the first arrived under a name
starting with "amp;". unescapeHtml moves into core beside escapeHtml and
runs before the trackable and skip-prefix checks.
