+++ packages/core/src/html.ts
@@ -8,6 +8,17 @@ export function escapeHtml(value: string): string {
     .replaceAll("'", "&#39;");
 }
 
+/** Inverse of escapeHtml: decode the five entities it emits. Ampersand last so
+ * `&amp;lt;` decodes to the literal `&lt;`, not `<`. */
+export function unescapeHtml(value: string): string {
+  return value
+    .replaceAll("&lt;", "<")
+    .replaceAll("&gt;", ">")
+    .replaceAll("&quot;", '"')
+    .replaceAll("&#39;", "'")
+    .replaceAll("&amp;", "&");
+}
+
 /**
  * Wordmark for system emails, hosted by the product site rather than the
  * instance: most self-hosted deployments sit on private or loopback hosts
